Renamed InitTraps to InitTrapsMACKE, so it wont make conflicts with Lhures InitTraps.
