+++ Game/GameMain/GameMain.vcxproj
@@ -120,7 +120,7 @@
     <ClCompile Include="gameSceneObjects\LootSystem.cpp" />
     <ClCompile Include="gameGraphics\TileRenderingPipeline.cpp" />
     <ClCompile Include="gameGraphics\SSAOPipeline.cpp" />
-    <ClCompile Include="gameTraps\InitTraps.cpp" />
+    <ClCompile Include="gameTraps\InitTrapsMACKE.cpp" />
     <ClCompile Include="gameTraps\TrapCollisionSystem.cpp" />
     <ClCompile Include="gameTraps\TrapSpawnerSystems.cpp" />
     <ClCompile Include="gameUtility\CameraEcsFunctions.cpp" />
@@ -186,7 +186,7 @@
     <ClInclude Include="gameSceneObjects\LootSystem.h" />
     <ClInclude Include="gameSceneObjects\SceneObjectComponents.h" />
     <ClInclude Include="gameSceneObjects\SceneObjectGlobals.h" />
-    <ClInclude Include="gameTraps\InitTraps.h" />
+    <ClInclude Include="gameTraps\InitTrapsMACKE.h" />
     <ClInclude Include="gameTraps\TrapComponents.h" />
     <ClInclude Include="gameTraps\TrapEvents.h" />
     <ClInclude Include="gameTraps\TrapSpawnerSystems.h" />
