+++ Game/GameMain/GameMain.vcxproj.filters
@@ -105,11 +105,11 @@
     <ClCompile Include="gameTraps\TrapSpawnerSystems.cpp">
       <Filter>Traps</Filter>
     </ClCompile>
-    <ClCompile Include="gameTraps\InitTraps.cpp">
-      <Filter>Init</Filter>
-    </ClCompile>
     <ClCompile Include="gameTraps\TrapCollisionSystem.cpp">
       <Filter>Traps</Filter>
+    </ClCompile>
+    <ClCompile Include="gameTraps\InitTrapsMACKE.cpp">
+      <Filter>Init</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -357,11 +357,11 @@
     <ClInclude Include="gameTraps\TrapSpawnerSystems.h">
       <Filter>Traps</Filter>
     </ClInclude>
-    <ClInclude Include="gameTraps\InitTraps.h">
-      <Filter>Init</Filter>
-    </ClInclude>
     <ClInclude Include="gameTraps\TrapCollisionSystem.h">
       <Filter>Traps</Filter>
+    </ClInclude>
+    <ClInclude Include="gameTraps\InitTrapsMACKE.h">
+      <Filter>Init</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
